# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API Test Backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/approval-stages": {
            "get": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Data Approval Stages",
                "description": "Mengambil Data Approval Stages",
                "operationId": "Get Data Approval Stages",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Approval Stages",
                                    "data": [
                                        {
                                            "id": 1,
                                            "approver_id": 1
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Create a new Approval Stages",
                "description": "Membuat data Approval Stages",
                "operationId": "Create Approval Stages",
                "requestBody": {
                    "description": "Approval Stagess data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "approver_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval Stagess created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "default": 1
                                        },
                                        "approver_id": {
                                            "type": "number",
                                            "default": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/approval-stages/{id}": {
            "put": {
                "tags": [
                    "Approval Stages"
                ],
                "summary": "Update a new Approval Stages",
                "description": "Membuat data Approval Stages",
                "operationId": "Update Approval Stages",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the approval stage",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Approval Stagess data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "approver_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval Stagess Updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "default": 1
                                        },
                                        "approver_id": {
                                            "type": "number",
                                            "default": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/approvers": {
            "post": {
                "tags": [
                    "Approvers"
                ],
                "summary": "Create a new Approver",
                "description": "Membuat data Approver",
                "operationId": "Create Approver",
                "requestBody": {
                    "description": "Approvers data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mariana"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approvers created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "default": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "default": "Mariana"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expense": {
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create a new expense",
                "description": "Membuat data expense",
                "operationId": "Create Expense",
                "requestBody": {
                    "description": "Expense data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expense/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Data Expense",
                "description": "Mengambil Data Expense",
                "operationId": "Get Data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Berhasil mengambil Data Expense",
                                    "data": [
                                        {
                                            "id": "1",
                                            "amount": 500000
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/expense/{id}/approve": {
            "patch": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update data approval expense",
                "description": "Update data approval expense",
                "operationId": "update approval Expense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Expense data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "amount": {
                                            "type": "integer"
                                        },
                                        "status_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    },
                    "403": {
                        "description": "Approval tidak sesuai urutan"
                    }
                }
            }
        }
    }
}